refactor makeColumnDefinition: extract exported function to modules
