+++ src/scripts/modules/gooddata-writer-v3/helpers/Constants.js
@@ -1,0 +1,21 @@
+import keyMirror from 'fbjs/lib/keyMirror';
+
+export const DataTypes = keyMirror({
+  BIGINT: null,
+  DATE: null,
+  DECIMAL: null,
+  INT: null,
+  VARCHAR: null
+});
+
+
+export const Types = keyMirror({
+  ATTRIBUTE: null,
+  CONNECTION_POINT: null,
+  DATE: null,
+  FACT: null,
+  HYPERLINK: null,
+  LABEL: null,
+  REFERENCE: null,
+  IGNORE: null
+});
