+++ src/scripts/modules/gooddata-writer-v3/helpers/columnsEditorDefinition.js
@@ -2,9 +2,9 @@ import {Map, fromJS} from 'immutable';
 import {parseParameters, createConfigParameters} from './rowParametersTable';
 import PreferencesHeader from '../react/components/PreferencesHeader';
 import PreferencesColumn from '../react/components/PreferencesColumn';
-import {prepareColumnContext, getInitialShowAdvanced} from './makeColumnDefinition';
 import makeColumnDefinition from './makeColumnDefinition';
-
+import prepareColumnContext from './prepareColumnContext';
+import getInitialShowAdvanced from './getInitialShowAdvanced';
 
 export default {
   isColumnValidFn: (column) => !makeColumnDefinition(column).getInvalidReason(),
