+++ src/scripts/modules/gooddata-writer-v3/helpers/getInitialShowAdvanced.js
@@ -1,0 +1,7 @@
+export default function getInitialShowAdvanced(columns) {
+  return !!columns.find(column => {
+    return !!(column.identifier ||
+              column.identifierLabel ||
+              column.identifierSortLabel);
+  });
+}
